requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.10
requests==2.31.0
python-dotenv==1.0.1

